pyspark-worker/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2